Remove grpcutil dialer
Grpcutil was created to return richer error information when the TLS
handshake failed. The grpc-go package has since received changes that
return the information. It is no longer necessary.

Replacing with an agent/client Dial helper that dials the SPIFFE server
with a standard gRPC dialer which is configured to do the proper SPIFFE
auth.

Signed-off-by: Andrew Harding <azdagron@gmail.com>