Updating telemetry plugin config examples
The syntax in the existing examples for statsd and dogstatsd config
is not valid if additional properties are added to the config objects.
The hcl unmarshaller will ignore all but the first property for an array
of objects when they are specified as follows:

telemetry {
    Statsd {
        address = "localhost:1337"
        new_property = "foo"
    }
    Statsd {
        address = "collector.example.org:8125"
        new_property = "bar"
    }
}

In the above example, "new_property" will be silently ignored by the hcl
parser and the corresponding values in the Go config objects will be empty.

The proper way to specify an array of objects in hcl appears to be as follows:

telemetry {
    Statsd = [
        { address = "localhost:1337" new_property = "foo" },
        { address = "collector.example.org:8125" new_property = "bar" },
    ]
}

This limitation of the array syntax definition is unfortunately not very
clearly documented anywhere, and there is an open issue on the GitHub page
requesting clarification of this syntax:
https://github.com/hashicorp/hcl/issues/276

Updating the documentation to be more future-proof in case new properties
are added to statsd and dogstatsd plugin configuration.

Signed-off-by: Ryan Turner <turner@uber.com>