Fix IsAlive() check on linux peertracker
IsAlive() current calls Readdir(1) which causes a false-negative when
the directory entries have all been enumerated io.EOF is returned.

The other issue is that the os.File object caches the directory entries,
so even if you Seek(0), the call still succeeds.

The solution is to use the syscall package directly to bypass caching.
Each call to syscall.ReadDirent will return entries until all have been
enumerated. At that point, it will return 0, which we still detect as
successful. As soon as the process is closed corresponding to the
/proc/pid directory being tracked, syscall.ReadDirent will return -1,
which IsAlive() interprets as failure.

Signed-off-by: Andrew Harding <azdagron@gmail.com>