Fix flaky minting tests
The fake server ca was accidentally initializing itself with a realtime
clock when a clock option was not provided. This caused assertion
failures verifying TTLs that were off by a second.

Fixed by initializing with a mock test. Removed the New() method from
the test clock package and fixed up other misuse to prevent this sort of
thing from happening again.

Signed-off-by: Andrew Harding <azdagron@gmail.com>