Speed up image building
Currently, SPIRE server and agent are built independently, running the
unit-test suite and building the alpine image twice. This change
consolidates all images into a single dockerfile and uses the `--target`
flag to indicate which image to target for each build. Due to the shared
multi-stage images, the unit-test suite and the alpine base image is
only run once. This cuts the time spent for the image
building/publishing stage in half.

Signed-off-by: Andrew Harding <azdagron@gmail.com>