Move build.sh global vars to local scope
The ARTIFACT_DIRS, RELEASE_DIRS, and SOURCE_PKGS vars are initialized
globally on every invocation of build.sh even though individually they
are used in one spot.

This change moves the initialization into functions that can be invoked
as needed instead.

This was breaking building the development container since it only pulls
in build.sh in order to run the setup step.

Signed-off-by: Andrew Harding <azdagron@gmail.com>