Add support for aws tags in agent ID
This change adds support for templating EC2 instance tags into
the agent base ID. There is also some refactoring that makes
this easier to reason about, and avoids having mutliple
code paths which create ec2 clients / fetch ec2 instances.

There is a test that verifies even if tags are used in the
agent ID but not supplied, the attestor will not panic.

Some minor changes were made to tests as a result of now
checking if an ec2 instance needs to be fetched earlier in
the process. This means the attestor is doing less wasted work.

Signed-off-by: Tyler Dixon <tylerd@uber.com>