Upgrade github.com/docker/docker
Fixes #1135

This was re-raised to me as a change that was breaking due to a somewhat edge case
use case in `pkg/agent/plugin/workloadattestor/docker/docker.go` where we were collecting
a slice of options to then pass to the constructor. docker/docker changed to have an exported
type for the options and the fix here is extremely simple. Additionally, its unlikely that
direct consumers of spiffe/spire are also doing this exact thing with docker/docker. If they
are, that is unfortunate but the fix is trivial and this should not hold spiffe/spire to an
ancient version of docker/docker.

Signed-off-by: Tyler Dixon <tylerd@uber.com>