Remove minikube-based k8s tests from v0.8 branch Somewhere along the way, the travis environment was updated in such a way that our minikube tests all broke. This breakage was observed by @azdagron to not be related to any changes on our side. A copy of the error is pasted below.
We knew that we wanted to move away from minikube for these tests in
favor of kind, so rather than shave the minikube yak, we rolled
forward. The result is that any commit prior to that change will no
longer pass the minikube-related tests in travis.

Unfortunately, the 0.8.2 release does not include kind support (which,
of course, the new test suite relies on). Rather than pull a new feature
into 0.8.2, this commit removes the minikube-related tests from
travis.yml in the v0.8 branch. This is justified because 1) we know the
tests to be passing, and 2) we expect this to be the last meaningful 0.8
release.

```
$ minikube update-context
*
X update config: Kubeconfig does not have a record of the machine cluster
*
* Sorry that minikube crashed. If this was unexpected, we would love to hear from you:
  - https://github.com/kubernetes/minikube/issues/new/choose
The command "minikube update-context" failed and exited with 70 during .
Your build has been stopped.
```

Signed-off-by: Evan Gilman <evan@scytale.io>