agent: Add timeout to server calls, clean up rotation logic
1. Add timeout to server calls. This enables the agent to fail more
predictably when the server is under load. The error is logged, which
provides some observability.

2. Clean up rotation logic. The ticker used in agent/manager ensures
there is always 5 seconds between invocations of synchronize(). A subtle
consequence of this, is that if a request takes 5 seconds, the ticker
immediately fires (since 5 seconds has elapsed since the last send to
the ticker channel) and synchronize() will be called immediately. This
change provides a little extra breathing room to the sync interval,
without adding any kind of backoff, which will be added in a subsequent
PR.

Signed-off-by: R. Tyler Julian <tjulian@uber.com>