Abbreviate git commit hash in version
Some metrics systems reject values that appear auto-generated.
Abbreviating the SHA1 Git commit hash in the version string is a
reasonable workaround for this. GitHub defaults to abbreviating the
commit hash to 7 characters in its UI in seemingly all repos it hosts.
For this reason, chose to abbreviate the hash to the first 7 characters
for easier correlation between SPIRE versions and the commit on which that
version is based.

Signed-off-by: Ryan Turner <turner@uber.com>