Fix kubernetes container restart race
The state show by the Kubelet api can lag the actual state of the system. For a newly created pod this was worked
around by polling until all containers in all pods on the system had a container ID, or until a timeout expired.

When a container within a pod is restarted, however, there's a window where the API reports the old container ID.
If the application in the new container starts rapidly enough then the agent will be unable to map the new ID to
a pod, and since it believes that all pods are "ready" it will immediately return no selectors.

This fix removes the !notAllContainersReady shortcut entirely. If a container ID cannot be looked up then we will
poll until MaxPollAttempts is hit, regardless of the reported state of containers on the system.

Signed-off-by: Jonathan Oddy <jonathan.oddy@transferwise.com>