Fix MySQL deadlock in SetNodeSelectors
SetNodeSelectors removes existing selector rows for a SPIFFE ID and
inserts new ones under a single transaction, essentially replacing the set of
node selectors for a SPIFFE ID. Unfortunately, with the default
transaction isolation level, this causes deadlocks on MySQL due to
aggressive gap locks on the index table. Even if the isolation level is
changed to "read committed", deadlocks still occur in the reattestation
case (i.e. node selector rows exist for a SPIFFE ID).

The fix is to change SetNodeSelectors to query for the rows to delete
and then issue the delete for the specific rows, which does not trigger
gap locks, fixing the deadlock.

Signed-off-by: Andrew Harding <azdagron@gmail.com>