Switch away from stateful sets for k8s integration test The k8s integration test fails often for unknown reasons. The failure is almost always while waiting for the SPIRE Server's database's stateful set.
This integration test does not exercise anything that depends on the
database being a stateful set - in fact, it does not exercise anything
that would require us use an external database. This commit switches
away from stateful set and postgres datastore in favor of sqlite3.
Hopefully this will cut down on the time it takes for the test suite to
run as well as improve the reliability of the suite as a whole.

Signed-off-by: Evan Gilman <evan@scytale.io>