Fix order of application in k8s integration test
Currently the webhook configuration is applied at the same time the
server and k8s-workload-registrar container is deployed. This causes
test flakiness if the control plane is configured to use the webhook
before the server pod is deployed, since the webhook obviously is not
available yet.

This change updates the k8s test to first apply all of the server and
registrar configuration, wait for the server pod to be fully deployed,
and then apply the webhook configuration.

Signed-off-by: Andrew Harding <azdagron@gmail.com>