Warn when unknown config options are encountered SPIRE recently gained a feature in which it can detect when unknown config options are present in its configuration file. When this condition is detected, an error message is returned and SPIRE exits. While this is probably the behavior that we want to have in the long run, it could be unexpected for a config that worked with 0.9.0 to stop working with 0.9.1.
To address this, it feels appropriate for us to first log a warning that
unknown config options have been detected. In a future ".0" release, we
can flip that behavior back to the current logic. Since 0.9.0 -> 0.10.0
is a supported upgrade path, and users of 0.9.0 don't receive this
warning, we should probably wait until 0.11.0 to do this.

Since the (current) plan is to eventually move this from a warning back
to bailing out, the warning message should probably state something to
that effect (e.g. "this will be a fatal error in a future release").
Once this PR is merged, I will send another with changes to the actual
logged messages.

Signed-off-by: Evan Gilman <evan@scytale.io>