Clean up catalog package layout
During the catalog refactor we needed to build plugin code
that itself relied on the catalog code. To keep the catalog package
clean, the test code was placed in a separate package. However this
introduced a cyclic import dependency. The approach taken at the time
was to move the common parts to an "internal" package that was consumed
by both catalog and the test package.

This change is a small cleanup that removes the need for the internal
package by using a distinct package for the unit test (go allows test
and non-test code in the same directory to have distinct packages),
breaking the import cycle.

Also removes some stutter (i.e. test.Test* => test.*) that causes golint
warnings.

Signed-off-by: Andrew Harding <azdagron@gmail.com>