Expose logrus options to `package main`
Switch log package to "Functional options" style
The functional options style of arguments is flexible as adding new options is
a non-breaking change to the NewLogger, and defaults are easier to use as those
options can be omitted.  Additionally they are less error-prone than a series
of positional string arguments which could easily be mixed up.

The options are plumbed up to cli.Run (in both server and agent).
Logrus has a number of options not exposed by SPIRE, and this allows those to
be configured in a custom `package main`.

I plan to use this for a custom Logrus formatter and hook to align with our
existing monitoring infrastructure.

Signed-off-by: Matthew McPherrin <mmc@squareup.com>