Fix autocert panic when cert key disappears
The GetPrivateKey method on the autocert KeyStore returns nil when the
key does not exist for example when a server configured with a memory
KeyManager is restarted. The autocert code did not properly handle this
case when loading the cert key, causing a panic.

This change changes the GetPrivateKey contract so that a special error
is returned if the key is missing to make it harder to mishandle. The
autocert code properly handles the error both when loading the account
key (by generating a new account key) and the cert key (treats it
as a cache miss).

Signed-off-by: Andrew Harding <azdagron@gmail.com>