Prune back CODEOWNERS file There are a few folks listed as CODEOWNERs in this repository who are no longer actively maintaining. This commit prunes the list back so that it reflects reality. In doing so, the "infrastructure maintenance" section became redundant so I simply removed it.
The following GitHub handles are affected:
* @drrt: Dave no longer maintains the SPIRE build machinery or
documentation. The core maintainers took up this responsibility some
time ago. I have spoken with Dave about this change, and he concurs.
* @jennybeth: Jenny no longer maintains the SPIRE documentation as she
has turned attention elsewhere. I spoke with Jenny about this, and she
indicated that she hoped to return to active maintenance in the future,
which I look forward to. We'll remove her from this list in the
meantime, and add her back as soon as she's ready.
* @martincapello: Martin no longer actively maintains SPIRE, and is
currently focused on other works. I spoke with him about this change,
and he concurs. We hope to get him back sometime soon!

Signed-off-by: Evan Gilman <evan@scytale.io>