Fix agent SVID rotation check interval
The agent SVID rotator has a bug whereby the rotation timer is extended
when a bundle update happens. This can cause the agent to fail to rotate
in time for small TTLs. It was made worse by a recent change to
introduce backoff to the rotation timer.

This change decouples the rotation timer and bundle update handling.

Signed-off-by: Andrew Harding <azdagron@gmail.com>