Prepare SQL statements for optimized queries
Recently optimized queries (ListRegistrationEntries,
FetchRegistrationEntry, GetNodeSelectors) are not being executed using
prepared statements causing significant, and unnecessary, database load.

This change introduces an on-demand statement cache to all but remove the
preparation cost for the queries. While it would be possible to
enumerate and prepare all the queries in advance, extra care would be
required to make sure all of the variants were considered. Although the
current approach isn't quite ideal, it is a step in the right direction
and requires no additional maintenance if more optimized queries are added.

Signed-off-by: Andrew Harding <azdagron@gmail.com>