Use AWS SDK to retrieve ec2 instance identity doc
This allows use of the v2 instance identity metadata service, which has
improved resilience to SSRF and similar vulnerabilities.

There is some backwards compatibility with the previous configuration that
specifies full URLs, however, the SDK wants you to configure a single endpoint.

This change is fully backwards compatible when using AWS IID or a direct proxy
for it.  It is theoretically incompatible if you used nonstandard URLs, but
that is unlikely given the response must be signed with a fixed Amazon
controlled private key.

Signed-off-by: Matthew McPherrin <mmc@squareup.com>