Use namespaced binary when calculating sha
The unix workload attestor can be configured to resolve the path and sha
as of the workload to selectors. It looks up the path to the running
process by following the `/proc/<pid>/exe` link, then uses that path
when calculating the sha.

When workloads are running in different namespaces (e.g. in docker), the
link resolved through `/proc` is the namespaced path to the workload -
if spire-agent is running in a different namespace to the workload, this
path can resolve to a _different_ file from the agent's perspective,
resulting in the sha of the wrong file being calculated.

Change the sha calculation to use `/proc/<pid>/exe` directly, rather
than the apparent path it links to, which will correctly resolve to the
namespaced workload.

Resolves spiffe/spire#1403

Signed-off-by: Ben Burry <benburry@github.com>