Remove PublishX509CA RPC
The UpstreamAuthority plugin was introduced as a solution to supported
JWT-SVID in nested SPIRE deployments. The intention of this RPC was to
support self-signed X.509 CAs in nested SPIRE architecture but that is a
change of much larger scope.

This change removes the PublishX509CA RPC to pare down amount of changes
needed in the core to adopt UpstreamAuthority. We can always add it back
later in a backwards compatible way.

Signed-off-by: Andrew Harding <azdagron@gmail.com>