Add Type to RegistrationEntry
There is currently no reliable way in SPIRE to distinguish
between node and workload registrations.
Introducing new Type property on a RegistrationEntry,
which has possible values of {UNKNOWN, NODE, WORKLOAD}
in order to provide this distinction.

Being able to distinguish between node and workload registrations will
enable at least two noticeable improvements:

1) Optimization of regentryutil/fetch algorithm to terminate recursive
   registration entry queries when the current registration is a
   workload registration (see Issue #980).
2) Ability to query for registrations by a specific type of
   registration. Most often when listing registrations, a client is only
   interested in one type of registration, node or workload. Performing
   this filtering on the query side is non-trivial due to the
   possibility of large registration entry sets and because the
   filtering logic today is highly based on consumer-specific SPIFFE ID
   path conventions.

The UNKNOWN Type value is used for any existing registrations which
do not have this Type field set. SPIRE Server will need to accept
CreateRegistrationEntry requests with the UNKNOWN Type value for one
additional release to maintain backwards compatibility.
In the N+1 release, SPIRE Server may disallow CreateRegistrationEntry
requests with the UNKNOWN Type value. When the client does not
explicitly set the Type field in a CreateRegistrationEntry request,
the Protobuf will resort to the default value of the
RegistrationEntryType enum, which is the value UNKNOWN.

Signed-off-by: Ryan Turner <turner@uber.com>