Disallow workload registrations from being a parent
Validate parent-child relationships in CreateEntry handler to
prevent workload entries from being the parent of any other entry.
Registrations can be created with a parent ID that does not exist as the
SPIFFE ID of an existing registration, so on each create registration
entry request evaluate whether the new registration would be a
parent if it is a workload registration.

Signed-off-by: Ryan Turner <turner@uber.com>