Implement an Informer strategy for k8s-workload-registrar
The registrar gains an option to use an informer instead of a
webhook. In this mode, it watches the k8s API instead of listening for
updates from a webhook.

The controller code is extended to react to "add" events by removing
any outdated entries for the same pod, so label changes are now
reflected in the registration entries.

Update client-go version to kubernetes 1.15 to get past API
changes. Regenerate mocks to match.

Signed-off-by: Andrew Suffield <asuffield@gmail.com>