Revert back to Go 1.13 Go 1.14 introduced asynchronous preemption, a feature which tickles a bug in certain versions of Linux. The Go team has put mitigations in place to try to work around this bug, but the end result still represents a sacrifice in both safety and reliability when running on affected systems.
The bug has since been patched, however the patch was recent and we feel
it will take some amount of time to trickle down - at least to an extent
at which we feel comfortable saying it's unlikely to be encountered.

The kernel bug affects everyone, not just us and not just Go. That being
said, this project is not receiving a significant gain from anything
that became available in Go 1.14, leaving risk without upside. As a
result, we've decided to downgrade to Go 1.13 for the time being.

We will re-evaluate the move to Go 1.14 in the near future.

Signed-off-by: Evan Gilman <evan@scytale.io>