Add CreateEntryIfNotExists RPC to Registration API It is often the case that registrars want to create an entry, or assert the presence of an entry, knowing that it may already exist. One example is if multiple registrar instances race against each other to create an entry. Another example is if a registrar loses track of what has or hasn't been registered, and wants to simply re-register everything.
This is mostly OK today, however it can be a bit noisy because we log an
error message whenever an error is returned via API, which the existing
CreateEntry call does. As we are currently looking to adopt this pattern
for our own k8s-workload-registrar code, it's important to make sure
that it (at the very least) _appears_ to be well supported!

The existing CreateEntry call does not have a dedicated request message,
making a new RPC required in order to address the problem. This commit
introduces the `CreateEntryIfNotExists` RPC, which behaves similarly to
the `CreateEntry` RPC with the exception that if the entry already
exists, the call will succeed. A boolean in the response message is
flipped when this is the case.

Signed-off-by: Evan Gilman <evan@scytale.io>