Stream bundle updates from UpstreamAuthority
The UpstreamAuthority plugin interface allows the plugin to stream
updated X.509 roots and JWT keys. This change updates the SPIRE CA
manager to service the streams returned from the MintX509CA and
PublishJWTKey RPCs. It appends the received material into the trust
bundle.

The UpstreamAuthority fake is also updated to facilitate testing this
functionality. Additionally, some small scaffolding is added for using
pregenerated keys in tests to cut down on runtime costs is introduced
(can be introduced into the codebase slowly).

Fixes #1451

Signed-off-by: Andrew Harding <azdagron@gmail.com>