Make validate and run share the same command flags
This uses the same default config and merging logic as run.  All flags from run
are now supported by validate.  This should make validate closer to the
real-deal.  Some minor refactoring in run was needed to export an appropriate
configuration-loading function, and a Help function.

While I was at it, `run` was updated to use an `env` instead of printing to
Stderr directly.  This was used by validate tests, but could be used for tests
against the run command in the future.

Signed-off-by: Matthew McPherrin <mmc@squareup.com>