api-refactor: SVID API service definition
This is the definition for the SVID API in the Server API refactor. This
is not set in stone and some decisions actively in discussion are sure
to change this, albeit superficially.

I'm introducing this now so we can kick off development. This API was a
good candidate since it was 1) very small and focused, 2) not likely to
change much as the API solidifies, 3) uses a broad range of SPIRE
internals to do its job. Developing this API will help establish
patterns and practices that should be applicable to other API
development.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>