Robust container ID parsing for K8S Workload attestor
The K8S Workload Attestor looks up the container ID for the workload by
examining the Kubernetes related cgroups for the process. The cgroup paths
have a lot of variety depending on the container runtime, QOS settings,
etc.

The existing code is too rigid in the what paths that it accepts. This change
updates the parsing code to accept a path that:
1) Has a kubepods segment
2) Followed by one or more path segments (for the pod UID, QOS settings,
etc.)
3) Trailing with the container ID

It accomodates cgroup names with the systemd suffixes (i.e., .slice,
.scope) as well as varied container runtimes (e.g., docker, crio).

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>