Authorization middleware scaffolding
This middleware can be used to set up authorization for the whole API
surface from a single spot.

The middleware provide information on the context for handlers, like
the caller address, SPIFFE ID, and X509-SVID.

This change also updates the rpccontext package to provide finer grained
values and remove specific details on how those values are determined.

It also removes the auth check in a bundle handler since the middleware is
now in charge of all authorization checks.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>