Authorization middleware
This middleware can be used to set up authorization for the whole API
surface from a single spot.

The authorizers provide information on the context for handlers, like
the SPIFFE ID, address, X509-SVID, downstream and admin entries, etc.

This change introduces the following authorizers:
- an admin caller authorizer
- an agent caller authorizer
- a downstream caller authorizer
- a local caller authorizer (i.e. unix domain socket)
- an "any" authorizer (used to annotate methods that don't require auth)
- an "any of" authorizer used to combine any of the above authorizers

It also removes the auth check in a bundle handler since the middleware is
now in charge of all authorization checks.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>