Fix "pid" config in nested-rotation integration test
docker-compose v2 allowed for pid sharing with a specific service,
however v3 only allows opting into the host pid namespace. A recent
update to docker-compose seems to no longer accept the v2 options, which
we were apparently using even though declaring our configuration as v3.

This fix updates the "pid" configuration to "host" for all of the
required containers and provides documentation on each instance for why
joining the host namespace is required.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>