Remove peertracker from server UDS
This change removes peertracker use on the server.

The server historically inspected the UID and GID of clients connecting
over UDS to determine if they belonged to the same user or group of the
server in order to grant access. At some point the server switched to
relying on the process-wide umask (`0027`) and a specific `chmod(0770)`
to restrict access to those running as as the same UID/GID of the
server.

The server now wholly relies on file permissions for access control to
the administrative UDS.

The peertracker code prevents tracks caller PIDs to mitigate PID re-use.
However, since the server is not relying on PID information for access
control, the peertracker provides no benefit. Furthermore, the
peertracker requires SPIRE server and UDS clients to share a process
namespace, which requires special (and often discourages) configuration
in containerized environments (e.g. `shareProcessNamespace` in K8s).

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>