Back fake datastore with sqlite3 datastore impl
The current datastore is a distinct in-memory implementation. The
semantics don't always line up with that returned by the SQL datastore.
This has caused subtle bugs that aren't caught by unit-tests.

This change rips out the fake datastore internals and replaces them with
the SQL datastore plugin using an in-memory SQLite3 database.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>