Move DataStore metrics out of SQL plugin
The metrics emitted by the SQL plugin are limited to simple call
counters over each RPC. This change moves the metrics emission out of
the plugin and into a generic DataStore wrapper that emits the call
metrics host-side.

This removes a great deal of mechanics around metric emission from the
plugin unittests, which was a maintenance smell. It also makes metrics
emission plugin agnostic.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>