Remove deprecated upstream_bundle configurable
Changes the server to always use the upstream bundle as the trust bundle
and removes the configurable.

One notably change in behavior is that the SPIRE server will no longer
present the self-signed CA with its own certificate during the TLS
handshake. The self-signed CA was added in the TLS handshake as a
side-effect of code added to support presenting the CA certificate
during the TLS handshake when an UpstreamAuthority had signed the CA
certificate but upstream_bundle was false (to allow agents to bootstrap
using the upstream bundle). That code is no longer required though since
the server will now always include the intermediate CA when signed by
upstream.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>