Provide details on why agent was denied permission
Agents can receive PermissionDenied for a variety of reasons. Current
feature requests (e.g. #1257, #1644) rely on the agent doing special
behavior based on why it was denied, for example, attempting
re-attestation if evicted (but not if banned).

This change introduces a new PermissionDeniedDetails message that
contains a reason enumeration that agents can use to determine why
permission was denied in order to respond appropriately.

It also fixes a missing agent authorization check in the new middleware
when an agent SVID has expired. Checking expiry at TLS handshake time is
insufficient because the SVID can expire after the handshake on long
lived connections that are still making requests.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>