Fix makefile to rebuild go-based tools
When the go version gets bumped, the tools built via go are not
rebuilt with the new version. This can cause subtle breakages like small
updates to the gofmt.

This change fixes the Makefile to properly rebuild the tools if they
aren't built with the correct version.

It also updates a few generated protobuf code files that changed due to
formatting with go1.14.4.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>