Use WithReturnConnectionError dial option
If there is a problem with the TLS handshake between the agent and
server (e.g. the bootstrap bundle is stale) the dialing operation
eventually fails when the dial timeout is hit (hard coded to 30s).
Unfortunately, details of the connection error are included in the
details of the returned error. All operators know is that the dial
operation timed out. This makes it hard to debug this types of failures.

grpc-go introduced a WithReturnConnectionError dial option a few months
back that provides helpful details about the connection error.

This change adds the dial option when the agent dials the server.

Upgrading grpc-go to v1.30.0 also required bumping the google API
dependency as well.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>