Don't attempt migrations on read-only replica
On first initialization of SPIRE Server against a new database cluster
with replication, SPIRE Server may attempt to create tables on a
read-only replica if `ro_connection_string` is configured in the `sql`
datastore plugin AND the read-write replica has not yet replicated the
new database/tables to the read-only replica.

Such write operations to a read-only replica will always fail, so do not
attempt them.

Signed-off-by: Ryan Turner <turner@uber.com>