Remove dependency on go-spiffe v1
This change removes dependencies on go-spiffe v1. It does not go all the
way to fully adopt all of the types and functionality of go-spiffe v2,
deferring that for a later time.

The impetus for this change is that both go-spiffe v1 and v2 have
distinct packages with the Workload API service definition. Importing both
in SPIRE causes conflicts registering the generated types work with gRPC.

At a future point we'll want to remove the old Workload API client code
under /api/workload and replace it with go-spiffe v2, as well as the
federation bundle endpoint client/server code.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>