Switch agent to use new server APIs
This change updates the agent to use the new server APIs
unconditionally and removes the experimental flag.

It is a little larger than expected, mostly due to updates to unit-tests
that had a strong coupling/dependency on the agent client, in particular
the manager tests. I was going to go for a larger refactor but that
would be mostly wasted effort since we have planned work that will shake
everything up in these layers quite a bit when we implement the agent
feature to do a "soft" restart when it needs to re-attest.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>