Fix entry conversion to include revision number
Revision number was left off of the conversion from new to old entry
type. This was not a bug at the time since the revision number is
ignored when supplied in a request.

However, other functions are starting to leverage the conversion
functions and having the conversion be lossy is problematic.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>