Fix rate limiter warning when handler not invoked
The rate limiter is supposed to warn if the handler did not invoke the
limiter (suggesting a bug in the handler implementation), except when
the handler returns InvalidArgument.

The conditional in the rate limiter middleware was incorrect causing the
warning to be issued even when the handler was never invoked.

This change fixes the rate limiter.

It also moves the authorization middleware above the rate limiting
middleware as a precaution.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>