Reimport the reconciling controller
This smushes the reconciling controller in alongside the new CRD mode.
It works enough to pass integration tests, but needs cleanup, and some fixes around how
connections are established.

Most of the feedback from the PR is incorporated in this change, other than closing the
various connections when main terminates. I'll fix that in a further commit as the merge
with the CRD mode changed how main operated significantly.

Signed-off-by: Jonathan Oddy <jonathan.oddy@transferwise.com>