Code version is not initialized in a pristine database.
The code version is not set in the migration table when the database is
newly initialized. If the server is shut down and a version with an
older schema is started, it will fail to start even if it is within the
+/- minor version compatibility boundary, since the "code version"
stored in the database will be effectively 0.0.0.

This change fixes the pristine database initialization code to
initialize the code version column appropriately.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>