Fix error handling in integration test scaffolding
The integration test scaffolding has fallen prey to bash-ism whereby
`set -e` is NOT honored in subshells, even if set explicitly. This has
caused integration tests to continue running when a command in a step script
fails when the exit status not explicitly tested, hiding test failures.

The fix is to NOT execute the test in a subshell, but rather invoke a
separate instance of bash against a copy of the step code with a
preamble containing the common functions which also sets up bash options
like errexit and pipefail.

The change also includes a fix to the nested rotation test which had a
failed invocation due to a missing docker-compose option disabling the
TTY.

There is also included a small self-test suite that ensures that command
failure fails step scripts. More tests can be added in the future that
ensure various mechanics of the integration test framework are working
as expected.

Signed-off-by: Andrew Harding <aharding@vmware.com>