Pregenerate a few more EC256 test keys
A future PR requires a few more EC256 keys than are currently
pregenerated. In regenerating new keys, I noticed that the standard
library RSA key generation is non-deterministic, even with a fixed
random source. This is "by design" to prevent folks from taking a
dependency on the key generation algorithm.... which is exactly what we
were trying to do to keep the churn down when changing the test key
count.

This change changes the generation code to simply lift in the existing
keys and removing/adding keys as necessary.

It is a little weird since it requires some tweaking of the generated
`keys.go` file when introducing a new key type before running the
generate.sh script, but considering how rare that event is, it's
probably OK?

This change also bumps the number of EC256 keys, since the future PR
needs them anyway. It also adds a missing cleanup step in the
generate.sh script to remove the temporary file on failure.

Signed-off-by: Andrew Harding <aharding@vmware.com>